dpdk-dev-libibverbs: allow meta-dpdk to play nice with other layers
If you are using meta-dpdk along with a layer which has its own
libibverbs recipe there is potential for conflict when assembling the
rootfs. Allow coexistence with "proper" libibverbs by making
dpdk-dev-libibverbs RPROVIDE "libibverbs1".

This can be seen for example when combining meta-dpdk and
meta-cloud-services with an image that includes dpdk and glusterfs.

(LOCAL REV: NOT UPSTREAM) Sent to meta-intel 19 Dec 2017

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>